uniformed controllers and repos

+++ src/main/java/dev/astranfalio/teioc/repository/AnswerRepository.java
@@ -3,9 +3,7 @@ package dev.astranfalio.teioc.repository;
 import dev.astranfalio.teioc.entity.AnswerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-// todo: check long with entity type
 @Repository
-public interface AnswerRepository extends JpaRepository<AnswerEntity, Long> {
+public interface AnswerRepository extends JpaRepository<AnswerEntity, Integer> {
 
 }
