+++ src/main/java/dev/astranfalio/teioc/repository/SurveyRepository.java
@@ -3,9 +3,7 @@ package dev.astranfalio.teioc.repository;
 import dev.astranfalio.teioc.entity.SurveyEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-// todo: check long with entity type
 @Repository
-public interface SurveyRepository extends JpaRepository<SurveyEntity, Long> {
+public interface SurveyRepository extends JpaRepository<SurveyEntity, Integer> {
 
 }
